decompose for better future
