+++ run.rb
@@ -1,5 +1,5 @@
 require_relative 'lib/bishop'
-
+require_relative 'lib/base'
 
 # use in this way - Bishop.new('b4').exec
